Reply (#22)
* feat : reply layout

* docs : 더미 데이터 추가

* docs : App.js 파일 수정

충돌이 나지 않도록 App.js 파일 수정했습니다

* feat : 파이어베이스 연동 완료, 답글 기능은 아직

* feat : 파이어베이스 연동 완료 답글 기능 아직

* docs : 필요없는 문서 삭제

* feat : 유저 정보 연동 완료'

* feat : 바뀐 데이터 구조 반영

* 파이어베이스 지수님 키로 바꿨습니다

파이어베이스 키 변동

* 파이어베이스 키 변경

파이어베이스 키 구교원 키로 변경

* docs: App.js 메인 페이지 수정

main 에 있는 App.js 코드로 수정

* 파이어베이스 키 변동

파이어베이스 키 변동

* feat : 답글 기능 추가

* docs : App.js 변경

* docs : firebase 문서 통일

* feat : 댓글에 닉네임, 비밀번호 기능 추가

+++ src/components/reply/index.jsx
@@ -2,7 +2,7 @@
 
 import React, {useState, useEffect} from 'react';
 import firebase from 'firebase';
-import db from '../../Firebase.js';
+import db from '../../firebase.js';
 import * as S from './styles.js';
 
 function Reply() {
@@ -11,19 +11,23 @@ function Reply() {
     let [postLiked, setPostLiked] = useState(false); // 글 좋아요 눌렀는지 여부
     let [replyLiked, setReplyLiked] = useState([false]);   // 댓글 좋아요 눌렀는지 여부
     let [replyArrow, setReplyArrow] = useState("↓");    // 댓글 버튼 화살표
-    let [replyAgain, setReplyAgain] = useState([false]); // 대댓글 쓰기 눌렀는지 여부
+    let [showReReply, setShowReReply] = useState([false]); // 대댓글 쓰기 눌렀는지 여부
     let [replyContent, setReplyContent] = useState(""); // 댓글 내용
+    let [reReplyContent, setReReplyContent] = useState(""); // 답글 내용
     let [replyEditContent, setReplyEditContent] = useState(""); // 댓글 수정 내용 저장
-    let [isEditing, setIsEditing] = useState([false]);  // 댓글을 수정 중인지
+    let [isEditing, setIsEditing] = useState([false]);  // 댓글을 수정 중인지 여부 저장
+    let [replyNickName, setReplyNickName] = useState();
+    let [replyPassword, setReplyPassword] = useState();
 
     // props로 주어지는 데이터
     let [categoryID, setCategoryID] = useState();
     let [postID, setPostID] = useState();
     
     // 정보 설정
-    let [post, setPost] = useState();
-    let [reply, setReply] = useState();
-    let [replyID, setReplyID] = useState();
+    let [post, setPost] = useState();   // 글 정보
+    let [reply, setReply] = useState(); // 댓글 정보
+    let [replyID, setReplyID] = useState(); // 댓글 아이디
+    let [reReplies, setReReplies] = useState(); // 답글 정보(이중 배열)
     
 
     // 유저 아이디, 글 아이디 설정
@@ -35,7 +39,7 @@ function Reply() {
 
 
     // 글 정보, 댓글 싹 다 가져오기
-    useEffect(()=>{
+    const getReplies = () => {
 
         if (categoryID !== undefined && postID !== undefined) {
             db.collection("categories").doc(categoryID)
@@ -57,21 +61,24 @@ function Reply() {
             postRef.onSnapshot((doc)=>{
                 setPost(doc.data());
             }); // 글 정보 가져오기
-        }  
+        }
+    }
+    useEffect(()=>{
+        getReplies();
     }, [categoryID, postID])
     
 
     // 댓글 개수를 바탕으로 필요한 상태들 초기화
-    useEffect(()=>{
-        
-        let tmpReplyAgain = [...replyAgain];
+
+    const initialize = () => {
+        let tmpShowReReply = [...showReReply];
         let tmpIsEditing = [...isEditing];
         let tmpReplyLiked = [...replyLiked];
 
         if (reply !== undefined) {            
             for (let i = 0; i < reply.length; i++) {
-                if (tmpReplyAgain[i] === undefined) {
-                    tmpReplyAgain[i] = false;
+                if (tmpShowReReply[i] === undefined) {
+                    tmpShowReReply[i] = false;
                 }
                 if (tmpIsEditing[i] === undefined) {
                     tmpIsEditing[i] = false;
@@ -80,10 +87,34 @@ function Reply() {
                     tmpReplyLiked[i] = false;
                 }
             }
-            setReplyAgain(tmpReplyAgain);
+            setShowReReply(tmpShowReReply);
             setIsEditing(tmpIsEditing);
             setReplyLiked(tmpReplyLiked);
         }
+    }
+    useEffect(()=>{
+        initialize();
+    }, [reply])
+
+
+    // 답글 가져오기
+    const getReReplies = () => {
+
+        if (reply) {
+
+            let tmpReReplies = [];
+            for (let i = 0; i < reply.length; i++) {
+                if (reply[i].reRepliesCnt > 0) {
+                    tmpReReplies.push(reply[i].reReplies);
+                } else {
+                    tmpReReplies.push([]);
+                }
+            }
+            setReReplies(tmpReReplies);
+        }
+    }
+    useEffect(()=>{
+        getReReplies();
     }, [reply])
 
 
@@ -122,12 +153,12 @@ function Reply() {
     }
 
 
-    // 대댓글 쓰는 창 보여주기
-    const showReplyAgain = i => {
+    // 답글 쓰는 창 보여주기
+    const clickReReply = i => {
 
-        let tmp = [...replyAgain];
+        let tmp = [...showReReply];
         tmp[i] = !tmp[i];
-        setReplyAgain(tmp);
+        setShowReReply(tmp);
     }
 
 
@@ -194,9 +225,16 @@ function Reply() {
 
     // 댓글 삭제하기
     const deleteReply = i => {
+
+        let password = prompt("삭제하시려면 비밀번호를 입력하세요");
+        if (password === reply[i].password) {
+        
         db.collection("categories").doc(categoryID)
         .collection("posts").doc(postID)
         .collection("replies").doc(replyID[i]).delete();
+        } else {
+            alert("비밀번호가 틀렸습니다");
+        }
     }
 
 
@@ -205,20 +243,75 @@ function Reply() {
 
         let today = new Date();
 
+        if (!replyNickName || !replyPassword) {
+            alert("닉네임, 비밀번호를 입력해주세요");
+        } else {
+            db.collection("categories")
+            .doc(categoryID).collection("posts")
+            .doc(postID).collection("replies").add(
+                {
+                    heart: 0,
+                    regDate: today.toLocaleDateString(),
+                    time: Date.now(),
+                    text: replyContent,
+                    reReplies: [],
+                    reRepliesCnt: 0,
+                    nickName: replyNickName,
+                    password: replyPassword
+                }
+            )
+            document.getElementById("replyContent").value = "";
+            document.getElementById("replyNickName").value = "";
+            document.getElementById("replyPassword").value = "";
+        }
+      }
+
+    
+    // 답글 전송
+    const submitReReply = i => {
+
+        if (reReplies[i]) {
+            var tmpReReplies = [...reReplies[i]];
+            tmpReReplies.push(reReplyContent);
+        } else{
+            var tmpReReplies = [reReplyContent];
+        }
+
         db.collection("categories")
         .doc(categoryID).collection("posts")
-        .doc(postID).collection("replies").add(
+        .doc(postID).collection("replies")
+        .doc(replyID[i]).update(
             {
-                heart: 0,
-                regDate: today.toLocaleDateString(),
-                time: Date.now(),
-                text: replyContent,
-                writer: "toodury"
+                reReplies : tmpReReplies,
+                reRepliesCnt : reply[i].reRepliesCnt + 1
             }
-        )
+        ) 
 
-        document.getElementById("replyContent").value = "";
-      }
+        document.getElementById("reReplyContent").value = "";
+
+        let tmpShowReReply = [...showReReply];
+        tmpShowReReply[i] = !tmpShowReReply[i];
+        setShowReReply(tmpShowReReply);
+    }
+
+
+    // 답글 삭제
+    const deleteReReply = (i, j) => {
+
+        let tmpReReplies = [...reReplies[i]];
+        tmpReReplies.splice(j, 1);
+
+        
+        db.collection("categories")
+        .doc(categoryID).collection("posts")
+        .doc(postID).collection("replies")
+        .doc(replyID[i]).update(
+            {
+                reReplies : tmpReReplies,
+                reRepliesCnt : reply[i].reRepliesCnt - 1
+            }
+        ) 
+    }
 
       
     return (
@@ -254,7 +347,7 @@ function Reply() {
                                             <>
                                             <S.Cancel onClick={()=>{cancelEditReply(i)}}>취소</S.Cancel>
                                             <S.WritingReply>
-                                                <S.WritingReplyUserName>작성자 이름</S.WritingReplyUserName>
+                                                <S.WritingReplyUserName>{reply[i].nickName}</S.WritingReplyUserName>
                                                 <S.WritingReplyInput defaultValue={reply[i].text} onChange={(e)=>{setReplyEditContent(e.target.value)}}></S.WritingReplyInput><br />
                                                 <S.ReplySubmitBtn onClick={()=>{editReply(i)}}>등록</S.ReplySubmitBtn>
                                             </S.WritingReply>
@@ -263,10 +356,10 @@ function Reply() {
                                     } else {
                                     return(
                                         <> 
-                                            <p>{reply[i].writer}</p>
+                                            <p>{reply[i].nickName}</p>
                                             <p dangerouslySetInnerHTML={{__html: reply[i].text}} />
                                             <S.ReplyDate>{reply[i].regDate}</S.ReplyDate>
-                                            <S.ReplyAgain onClick={()=>{showReplyAgain(i)}}>답글</S.ReplyAgain>
+                                            <S.ReplyAgain onClick={()=>{clickReReply(i)}}>답글</S.ReplyAgain>
                                             <S.ReplyDelete onClick={()=>deleteReply(i)}>삭제</S.ReplyDelete>
                                             <S.ReplyEdit onClick={()=>{showEditReply(i)}}>수정</S.ReplyEdit>
                                             <p><S.ReplyLikeBtn onClick={()=>clickReplyLikeButton(i)}>
@@ -277,11 +370,24 @@ function Reply() {
                                             }
                                             </S.ReplyLikeBtn>  {reply[i].heart}</p>
                                             {
-                                                replyAgain[i]
+                                                reply[i].reReplies
+                                                ? <>{
+                                                    reply[i].reReplies.map((reReply, j)=>{
+                                                        return (
+                                                            <S.ReReply>
+                                                                ➡{reReply}
+                                                                <S.ReplyDelete onClick={()=>deleteReReply(i, j)}>삭제</S.ReplyDelete>
+                                                            </S.ReReply>
+                                                        )
+                                                    })
+                                                }</>
+                                                : null
+                                            }
+                                            {
+                                                showReReply[i]
                                                 ?   <S.WritingReply>
-                                                        <S.WritingReplyUserName>작성자 이름</S.WritingReplyUserName>
-                                                        <S.WritingReplyInput placeholder="답글 쓰기"></S.WritingReplyInput><br />
-                                                        <S.ReplySubmitBtn>등록</S.ReplySubmitBtn>
+                                                        <S.WritingReplyInput id="reReplyContent" placeholder="답글 쓰기" onChange={(e)=>{setReReplyContent(e.target.value)}}></S.WritingReplyInput><br />
+                                                        <S.ReplySubmitBtn onClick={()=>submitReReply(i)}>등록</S.ReplySubmitBtn>
                                                     </S.WritingReply>
                                                 : null
                                             }
@@ -292,8 +398,9 @@ function Reply() {
                             }
                             <p>
                                 <S.WritingReply>
-                                    <S.WritingReplyUserName>작성자 이름</S.WritingReplyUserName>
-                                    <S.WritingReplyInput id="replyContent" placeholder="댓글 쓰기" onChange={(e)=>{setReplyContent(e.target.value)}}></S.WritingReplyInput><br />
+                                    <S.WritingReplyNickName type="text" id="replyNickName" placeholder="닉네임" onChange={(e)=>{setReplyNickName(e.target.value)}}></S.WritingReplyNickName>
+                                    <S.WritingReplyPassword type="password" id="replyPassword" placeholder="비밀번호" onChange={(e)=>{setReplyPassword(e.target.value)}}></S.WritingReplyPassword>
+                                    <S.WritingReplyInput id="replyContent" placeholder="댓글 쓰기" onChange={(e)=>{setReplyContent(e.target.value)}}></S.WritingReplyInput><br/>
                                     <S.ReplySubmitBtn onClick={submitReply}>등록</S.ReplySubmitBtn>
                                 </S.WritingReply>
                             </p>
